INF339 Bases de Datos a Gran Escala/Tareas/Tarea 3/workspace_hrl copy/workspace_hrl copy/requirements.txt
fastavro
apache-beam
apache-airflow==2.9.3
pyarrow
pandas
python-snappy
confluent-kafka
minio
boto3
Faker
redis
psycopg
requests